Extract KAPI HVPA reconciliation into dedicated file
